+++ react/ReviewForm.tsx
@@ -1,17 +1,9 @@
 /* eslint-disable no-console */
-import React, { FC, Fragment, useContext, useEffect, useReducer } from 'react'
-import { ProductContext } from 'vtex.product-context'
-import ApolloClient, { ApolloQueryResult } from 'apollo-client'
-import { NormalizedCacheObject } from 'apollo-cache-inmemory'
-import { withApollo } from 'react-apollo'
-import {
-  FormattedMessage,
-  InjectedIntlProps,
-  injectIntl,
-  defineMessages,
-} from 'react-intl'
-import flowRight from 'lodash.flowright'
-import path from 'ramda/es/path'
+import React, { Fragment, useEffect, useReducer } from 'react'
+import { useProduct } from 'vtex.product-context'
+import { ApolloQueryResult } from 'apollo-client'
+import { useApolloClient } from 'react-apollo'
+import { FormattedMessage, defineMessages, useIntl } from 'react-intl'
 import { useCssHandles } from 'vtex.css-handles'
 import { Card, Input, Button, Textarea } from 'vtex.styleguide'
 
@@ -20,21 +12,11 @@ import NewReview from '../graphql/newReview.graphql'
 import HasShopperReviewed from '../graphql/hasShopperReviewed.graphql'
 import StarPicker from './components/StarPicker'
 
-interface Product {
-  productId: string
-  productName: string
-}
-
 interface AppSettings {
   allowAnonymousReviews: boolean
   requireApproval: boolean
   useLocation: boolean
   defaultOpen: boolean
-}
-
-interface Props {
-  client: ApolloClient<NormalizedCacheObject>
-  settings?: AppSettings
 }
 
 interface HasShopperReviewedData {
@@ -212,15 +194,14 @@ const messages = defineMessages({
 
 const CSS_HANDLES = ['formContainer'] as const
 
-export const ReviewForm: FC<InjectedIntlProps & Props> = ({
-  intl,
-  client,
-  settings,
-}) => {
+export function ReviewForm({ settings }: { settings?: Partial<AppSettings> }) {
+  const client = useApolloClient()
+  const intl = useIntl()
+
   const handles = useCssHandles(CSS_HANDLES)
 
-  const { product } = useContext(ProductContext) as any
-  const { productId }: Product = product || {}
+  const { product } = useProduct() ?? {}
+  const { productId } = product ?? {}
 
   const [state, dispatch] = useReducer(reducer, initialState)
 
@@ -237,10 +218,8 @@ export const ReviewForm: FC<InjectedIntlProps & Props> = ({
       }
 
       const { namespaces } = sessionRespose
-      const storeUserId = path(
-        ['authentication', 'storeUserId', 'value'],
-        namespaces
-      )
+      const storeUserId = namespaces?.authentication?.storeUserId?.value
+
       if (!storeUserId) {
         return
       }
@@ -252,8 +231,8 @@ export const ReviewForm: FC<InjectedIntlProps & Props> = ({
 
       const profile = {
         email:
-          path(['email', 'value'], namespaces.profile) ||
-          path(['storeUserEmail', 'value'], namespaces.authentication),
+          namespaces.profile?.email?.value ??
+          namespaces.authentication?.storeUserEmail?.value,
       }
 
       if (typeof profile.email !== 'string') {
@@ -508,4 +487,4 @@ export const ReviewForm: FC<InjectedIntlProps & Props> = ({
   )
 }
 
-export default flowRight([withApollo, injectIntl])(ReviewForm)
+export default ReviewForm