+++ react/admin/components/util/reviewCellRenderer.tsx
@@ -2,12 +2,7 @@
 /* eslint-disable jsx-a11y/click-events-have-key-events */
 import React, { useState } from 'react'
 import copy from 'copy-to-clipboard'
-import {
-  FormattedMessage,
-  defineMessages,
-  injectIntl,
-  InjectedIntlProps,
-} from 'react-intl'
+import { FormattedMessage, defineMessages, useIntl } from 'react-intl'
 import { IconCopy, ToastConsumer, Tooltip } from 'vtex.styleguide'
 
 import Stars from '../../../components/Stars'
@@ -18,7 +13,7 @@ const DEFAULT_TOAST_DURATION_MS = 1500
 
 type Props = {
   cellData: ReviewSummary
-} & InjectedIntlProps
+}
 
 const messages = defineMessages({
   copyTitle: {
@@ -31,8 +26,9 @@ const messages = defineMessages({
   },
 })
 
-const IntlReviewCellRenderer: React.FC<Props> = ({ cellData, intl }) => {
+const ReviewCellRenderer: React.FC<Props> = ({ cellData }) => {
   const [showCopy, setShowCopy] = useState(false)
+  const intl = useIntl()
 
   return cellData ? (
     <ToastConsumer>
@@ -94,7 +90,5 @@ const IntlReviewCellRenderer: React.FC<Props> = ({ cellData, intl }) => {
     </ToastConsumer>
   ) : null
 }
-
-const ReviewCellRenderer = injectIntl(IntlReviewCellRenderer)
 
 export default ReviewCellRenderer