+++ react/typings/vtex.render-runtime.d.ts
@@ -1,5 +1,0 @@
-import 'vtex.render-runtime'
-
-declare module 'vtex.render-runtime' {
-  export const canUseDOM: boolean
-}