+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- App types.
+- DOM hierarchy.
+
+### Changed
+- Use Apollo and React Intl hooks instead of HOCs.
 
 ## [2.3.1] - 2020-12-14